requirements.txt
fastapi==0.115.12
bcrypt==3.2.0
passlib==1.7.4
pydantic==2.11.1
pydantic-settings==2.9.1
SQLAlchemy==2.0.40
uvicorn==0.34.0
asyncpg==0.30.0
PyJWT==2.10.1
pydantic[email]
python-multipart
geoalchemy2
cryptography
shapely